feat: remove decideContainerOverrides
