+++ dotcom-rendering/.storybook/decorators/splitThemeDecorator.tsx
@@ -147,6 +147,7 @@ export const splitTheme =
 				data-color-scheme="light"
 				css={[
 					css`
+						color-scheme: light;
 						background-color: ${sourcePalette.neutral[100]};
 						color: ${sourcePalette.neutral[0]};
 					`,
@@ -171,6 +172,7 @@ export const splitTheme =
 				data-color-scheme="dark"
 				css={[
 					css`
+						color-scheme: dark;
 						background-color: ${sourcePalette.neutral[0]};
 						color: ${sourcePalette.neutral[100]};
 					`,
